fix: Moved the Skyroot Sign WoodType to be registered earlier on. This should ***hopefully*** fix the bug where instances of the mod randomly don't have a working Skyroot Sign texture.
