+++ src/main/java/com/gildedgames/aether/common/block/dungeon/ChestMimicBlock.java
@@ -31,7 +31,6 @@ import net.minecraft.core.Direction;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class ChestMimicBlock extends BaseEntityBlock implements SimpleWaterloggedBlock
@@ -40,7 +39,7 @@ public class ChestMimicBlock extends BaseEntityBlock implements SimpleWaterlogge
 	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 	protected static final VoxelShape AABB = Block.box(1.0D, 0.0D, 1.0D, 15.0D, 14.0D, 15.0D);
 
-	public ChestMimicBlock(BlockBehaviour.Properties properties) {
+	public ChestMimicBlock(Properties properties) {
 		super(properties);
 		this.registerDefaultState(this.stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(WATERLOGGED, Boolean.FALSE));
 	}
