+++ src/main/java/com/gildedgames/aether/common/block/entity/SkyrootSignBlockEntity.java
@@ -1,4 +1,4 @@
-package com.gildedgames.aether.common.entity.tile;
+package com.gildedgames.aether.common.block.entity;
 
 import com.gildedgames.aether.common.registry.AetherBlockEntityTypes;
 import net.minecraft.core.BlockPos;
@@ -6,14 +6,17 @@ import net.minecraft.world.level.block.entity.SignBlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class SkyrootSignTileEntity extends SignBlockEntity
+import javax.annotation.Nonnull;
+
+public class SkyrootSignBlockEntity extends SignBlockEntity
 {
-    public SkyrootSignTileEntity(BlockPos p_155700_, BlockState p_155701_) {
+    public SkyrootSignBlockEntity(BlockPos p_155700_, BlockState p_155701_) {
         super(p_155700_, p_155701_);
     }
 
+    @Nonnull
     @Override
-    public BlockEntityType<SkyrootSignTileEntity> getType()
+    public BlockEntityType<SkyrootSignBlockEntity> getType()
     {
         return AetherBlockEntityTypes.SKYROOT_SIGN.get();
     }
