+++ src/main/java/com/gildedgames/aether/common/block/utility/IncubatorBlock.java
@@ -27,8 +27,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.level.Level;
 
-import net.minecraft.world.level.block.state.BlockBehaviour;
-
 import javax.annotation.Nonnull;
 import java.util.Random;
 
@@ -36,7 +34,7 @@ public class IncubatorBlock extends BaseEntityBlock
 {
 	public static final BooleanProperty LIT = BlockStateProperties.LIT;
 
-	public IncubatorBlock(BlockBehaviour.Properties builder) {
+	public IncubatorBlock(Properties builder) {
 		super(builder);
 		this.registerDefaultState(this.stateDefinition.any().setValue(LIT, false));
 	}
